Added other tests to cmake.
